% Generated by roxygen2 (4.0.1): do not edit by hand
\name{download_data}
\alias{download_data}
\title{Download SuBLIME data}
\usage{
download_data(folder = system.file(package = "SuBLIME"), force = FALSE)
}
\arguments{
\item{download_data}{}

\item{folder}{Folder to download the data - usually SuBLIME folder,
but may need a different directory due to permissions}

\item{force}{Force download of file even if it exists}
}
\value{
Indicator if the file was downloaded and unzipped
}
\description{
Download test data for examples
}

