% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sublime_model.R
\docType{data}
\name{flairt1_sublime_model}
\alias{flairt1_sublime_model}
\title{SuBLIME Predictive model with only T1 and FLAIR}
\format{An \code{glm} object, but with data an other things, notably 
qr removed}
\usage{
flairt1_sublime_model
}
\description{
Predictive model for SuBLIME algorithm with only T1 and FLAIR.
NOTE: this may perform much worse than the original SuBLIME model
}
\keyword{datasets}
