% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Intensity-normalization by a image mask}
\usage{
normalize(image, mask = NULL)
}
\arguments{
\item{image}{3D Array or object of class nifti}

\item{mask}{Mask with same dimensions.  If NULL, then original image is returned}
}
\value{
Object of class nifti or array, depending on image input
}
\description{
This function normalized the image by the mean and
standard deviation by intensities of voxels in the mask.  In SuBLIME
the mask is normal appearing white matter
}
\examples{
\dontrun{
## put in ex here
}
}
\keyword{normalize}

