\name{SuBLIME-package}
\alias{SuBLIME-package}
\alias{SuBLIME}
\docType{package}
\title{ SuBLIME package: Subtraction-Based Logistic Inference for Modeling and Estimation}
\description{
This packages runs the SuBLIME pipeline on preprocessed data for detection of lesions in patients iwth MS.  The functions will normalize and runt he model
}
\details{
\tabular{ll}{
Package: \tab SuBLIME\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-08-12\cr
License: \tab What license is it under?\cr
}

}
\author{
Elizabeth Sweeney, John Muschelli 

Elizabeth M. Sweeney <emsweene@jhsph.edu>
}
\references{
Sweeney, E. M., Shinohara, R. T., Shea, C. D., Reich, D. S., & Crainiceanu, C. M. (2013). Automatic Lesion Incidence Estimation and Detection in Multiple Sclerosis Using Multisequence Longitudinal MRI. American Journal of Neuroradiology, 34(1), 68-73.
}
\keyword{ package }
\seealso{
~~ \code{\link[AnalyzeFMRI:GaussSmoothArray]{AnalyzeFMRI}} ~~
}
