% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sublime_model.R
\docType{data}
\name{sublime_model}
\alias{sublime_model}
\title{SuBLIME Predictive model}
\format{An \code{lm} object, but with data an other things, notably 
qr removed}
\usage{
sublime_model
}
\description{
Predictive model for SuBLIME algorithm
}
\references{
Sweeney, E. M., et al. "Automatic lesion incidence estimation and detection in multiple sclerosis using multisequence longitudinal MRI." American Journal of Neuroradiology 34.1 (2013): 68-73.
}
\keyword{datasets}
