% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sublime_model.R
\docType{data}
\name{nopd_sublime_model}
\alias{nopd_sublime_model}
\title{SuBLIME Predictive model without PD}
\format{An \code{glm} object, but with data an other things, notably 
qr removed}
\usage{
nopd_sublime_model
}
\description{
Predictive model for SuBLIME algorithm without PD modality
NOTE: this may perform much worse than the original SuBLIME model
}
\keyword{datasets}
